requirements.txt
fastapi
uvicorn
openai
llama-index
qdrant-client
python-dotenv
ragas
GitPython